+++ CHANGELOG.md
@@ -7,6 +7,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2] - 2023-06-26
+
+### Added
+
+- Skip to content link
+
+### Changed
+
+- Updated docs
+- Updated code of conduct
+- Fixed command dialog rendered twice bug
+- Upgraded packages
+
+### Removed
+
+- Avatar component
+
+## [0.3.1] - 2023-06-20
+
+### Added
+
+- Accent color
+- Issue templates
+
+### Changed
+
+- Updated usage docs
+
 ## [0.3.0] - 2023-06-18
 
 ### Added